Remove misused "provably"
The word "provably" is almost always misued by computer scientists. To
detect such misuse ask yourself whether the meaning changes when you
delete the word "provably". If it doesn't, then you misused the word.
Here's an instance: "Mergesort runs provably in O(n * log n)" vs.
"Mergesort run in O(n * log n)". The running time of mergesort has
nothing to do with anybody having proved that -- it just **is** a fact
that it runs in O(n * log n), and it will do so even if no human ever
proved this fact.