Change representation of spines
This is a major change in the abstract syntax. Our representation of
spines was confused because the bound variables in the spine typing
information were bound in the arguments of the spine. I.e, given a spine

   e e1 e2

with typing annotation (x1:t1) (x2:t2) . t, the variable x1 was bound
in e2, but it should only be bound in t2. This of course is confusing
at least.

The benefit of the previous presentation was that there could be no
mismatch between the number of arguments and the number of binders in
the typing annotation. The new representation allows us to create an
invalid spine in which there are a different number of arguments and
binders.

Nevertheless, the code seems to be a bit more logical with the new
change.