General hints might not have a key (RFC)
Any of a hint's type, lhs and rhs may be a pattern variable, so when we
encounter one we give up and return None. This means that we store
general hints in two stores: a map over the keys and a list. When
looking them up a hint the map is preferred but the list is always
added.

Another option would be to just use a list for general hints and filter
when a key is available. Filtering would be slower than map look-up, but
we could guarantee that the hints are still returned in the same order
they were installed.