RFC: normalisation of hints
When we try to apply a hint to a term, the term is brought into whnf.
The hint may not apply unless it is also in whnf, but we currently just
install them as they come. The whnf command is not sufficient as we need
to whnf the components (lhs and rhs) of the type of a hint.

Should we normalise the components of hints when we create them? If
there is no principled solution we could raise an operation that allows
the user to normalise in a given way (probably too cumbersome).