Rename modules to structures
If we target Andromeda at formalization of mathematics then it makes
most sense to use mathematical terminology, in which case signatures go
with structures. For instance, we have a signature for a group, while a
group is a structure. I know this is not the only possibility, but
"module" means something completely different in mathematics and we
should avoid it.

Other changes:
* make pattern keys for signatures and structures a bit more specific
* remove typing annotations in fields as they are not needed
* organize the tokens in `parser.mly`
*