New emacs mode, eri as default indentation engine
The smie indentation works alright for the ML fragment but not for the
full grammar, so instead we default to using the eri.el
library ("enhanced relative indentation"). This choice can be customised
by setting `m31-indent-smart'.

Keybindings provided for eri are <tab> to increase and <shift-tab> to
decrease indentation.