Reorganization of syntax (WIP).
This is a partial commit which does not compile. But the plan is
to create more intermediate langauges:

* `Input` is produced by the parser and consumed by desugaring
* `Dsyntax` is produced by desugaring and consumed by typechecking
* `Rsyntax` is produced by typechekcing and consumed by runtime

This is necessary because with fancier typing annotations we cannot get
away with the polymorphic `Syntax` anymore. And that was ugly as hell
anyhow, and had no chance of surviving for long.

At the same time this code is implementing `let`-annotations (they
caused the split of syntax levels).