Remove printing of types of terms.
Printing the type of a term by calculating the type on-the-fly during printing
is very annoying to fix. We cannot simply just start printing an abstraction
`{x : A} e` and calculate the type of `e` when we get to it, because `e` may
refer to the bound variable `x`. So, if we really wanted to do this, we'd have
to unabstract as we print and so on. It's just not worth doing.

If we ever really want to print types of terms automatically, we should provide
them as input to `TT.print_term` or some such. Another, probably better option
is for `Jdg.is_term` to simply store its boundary (the type). This may have some
nice effects, for instance we never have to recalculate the type (at the expense
of using more memory).