Tag pattern meta-variables with questionmarks.
The current way of treating pattern variables in not viable.
Suppose a module `M` uses somewhere a pattern variable `A`.

If we run

1. `require M`
2. `rule A type`

then it works as expected, but if we run

1. `rule A type`
2. `require M`

then `A` inside `M` is treated as a type constructor, and no warning or error is ussued immediately.

The solution is to write `?A` in a pattern.